# csharp: improve performance and reduce memory footprint when creating and verifying webhook signatures