--- a/csharp/Svix/Svix.csproj
+++ b/csharp/Svix/Svix.csproj
@@ -7,6 +7,7 @@
     <Authors>Svix</Authors>
     <Company>Svix</Company>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <LangVersion>13</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>