--- a/csharp/Svix/Utils.cs
+++ b/csharp/Svix/Utils.cs
@@ -5,13 +5,6 @@ namespace Svix
 {
     internal static class Utils
     {
-        public static bool SecureCompare(string a, string b)
-        {
-            ArgumentNullException.ThrowIfNull(a);
-            ArgumentNullException.ThrowIfNull(b);
-            return SecureCompare(a.AsSpan(), b.AsSpan());
-        }
-        
         // Borrowed from Stripe-dotnet https://github.com/stripe/stripe-dotnet/blob/7b62c461d7c0cf2c9e06dce5e564b374a9d232e0/src/Stripe.net/Infrastructure/StringUtils.cs#L30
         // basically identical to SecureCompare from Rails::ActiveSupport used in our ruby lib
         [MethodImpl(MethodImplOptions.NoOptimization)]