--- a/csharp/Svix/Webhook.cs
+++ b/csharp/Svix/Webhook.cs
@@ -1,5 +1,7 @@
 ﻿using Svix.Exceptions;
 using System;
+using System.Buffers;
+using System.Buffers.Text;
 using System.Net;
 using System.Security.Cryptography;
 using System.Text;
@@ -17,14 +19,18 @@ namespace Svix
         internal const string UNBRANDED_SIGNATURE_HEADER_KEY = "webhook-signature";
         internal const string UNBRANDED_TIMESTAMP_HEADER_KEY = "webhook-timestamp";
 
+        private const int SIGNATURE_LENGTH_BYTES = 32;
+        private const int SIGNATURE_LENGTH_BASE64 = 48;
+        private const int SIGNATURE_LENGTH_STRING = 56;
         private const int TOLERANCE_IN_SECONDS = 60 * 5;
-        private static string prefix = "whsec_";
-        private byte[] key;
+        private const string PREFIX = "whsec_";
+        
+        private readonly byte[] key;
         public Webhook(string key)
         {
-            if (key.StartsWith(prefix))
+            if (key.StartsWith(PREFIX))
             {
-                key = key.Substring(prefix.Length);
+                key = key.Substring(PREFIX.Length);
             }
 
             this.key = Convert.FromBase64String(key);
@@ -35,7 +41,7 @@ namespace Svix
             this.key = key;
         }
 
-        public void Verify(string payload, WebHeaderCollection headers)
+        public void Verify(ReadOnlySpan<char> payload, WebHeaderCollection headers)
         {
             ArgumentNullException.ThrowIfNull(headers);
             
@@ -47,16 +53,16 @@ namespace Svix
             ArgumentNullException.ThrowIfNull(payload);
             ArgumentNullException.ThrowIfNull(headersProvider);
             
-            string msgId = headersProvider(SVIX_ID_HEADER_KEY);
-            string msgSignature = headersProvider(SVIX_SIGNATURE_HEADER_KEY);
-            string msgTimestamp = headersProvider(SVIX_TIMESTAMP_HEADER_KEY);
+            ReadOnlySpan<char> msgId = headersProvider(SVIX_ID_HEADER_KEY);
+            ReadOnlySpan<char> msgSignature = headersProvider(SVIX_SIGNATURE_HEADER_KEY);
+            ReadOnlySpan<char> msgTimestamp = headersProvider(SVIX_TIMESTAMP_HEADER_KEY);
 
-            if (String.IsNullOrEmpty(msgId) || String.IsNullOrEmpty(msgSignature) || String.IsNullOrEmpty(msgTimestamp))
+            if (msgId.IsEmpty || msgSignature.IsEmpty || msgTimestamp.IsEmpty)
             {
                 msgId = headersProvider(UNBRANDED_ID_HEADER_KEY);
                 msgSignature = headersProvider(UNBRANDED_SIGNATURE_HEADER_KEY);
                 msgTimestamp = headersProvider(UNBRANDED_TIMESTAMP_HEADER_KEY);
-                if (String.IsNullOrEmpty(msgId) || String.IsNullOrEmpty(msgSignature) || String.IsNullOrEmpty(msgTimestamp))
+                if (msgId.IsEmpty || msgSignature.IsEmpty || msgTimestamp.IsEmpty)
                 {
                     throw new WebhookVerificationException("Missing Required Headers");
                 }
@@ -64,34 +70,42 @@ namespace Svix
 
             var timestamp = Webhook.VerifyTimestamp(msgTimestamp);
 
-            var signature = this.Sign(msgId, timestamp, payload);
-            var expectedSignature = signature.Split(',')[1];
+            Span<char> expectedSignature = stackalloc char[SIGNATURE_LENGTH_STRING];
+            CalculateSignature(msgId, timestamp, payload, expectedSignature, out var charsWritten);
+            expectedSignature = expectedSignature.Slice(0, charsWritten);
 
-            var passedSignatures = msgSignature.Split(' ');
-            foreach (string versionedSignature in passedSignatures)
+            var signaturePtr = msgSignature;
+            var spaceIndex = signaturePtr.IndexOf(' ');
+            do
             {
-                var parts = versionedSignature.Split(',');
-                if (parts.Length < 2)
+                var versionedSignature = spaceIndex < 0 
+                    ? msgSignature : signaturePtr.Slice(0, spaceIndex);
+                
+                signaturePtr = signaturePtr.Slice(spaceIndex + 1);
+                spaceIndex = signaturePtr.IndexOf(' ');
+                
+                var commaIndex = versionedSignature.IndexOf(',');
+                if (commaIndex < 0)
                 {
                     throw new WebhookVerificationException("Invalid Signature Headers");
                 }
-                var version = parts[0];
-                var passedSignature = parts[1];
-
-                if (version != "v1")
+                var version = versionedSignature.Slice(0, commaIndex);
+                if (!version.Equals("v1", StringComparison.InvariantCulture))
                 {
                     continue;
                 }
+                var passedSignature = versionedSignature.Slice(commaIndex + 1);
                 if (Utils.SecureCompare(expectedSignature, passedSignature))
                 {
                     return;
                 }
-
             }
+            while(spaceIndex >= 0);
+            
             throw new WebhookVerificationException("No matching signature found");
         }
 
-        private static DateTimeOffset VerifyTimestamp(string timestampHeader)
+        private static DateTimeOffset VerifyTimestamp(ReadOnlySpan<char> timestampHeader)
         {
             DateTimeOffset timestamp;
             var now = DateTimeOffset.UtcNow;
@@ -105,26 +119,67 @@ namespace Svix
                 throw new WebhookVerificationException("Invalid Signature Headers");
             }
 
-            if (timestamp < (now.AddSeconds(-1 * TOLERANCE_IN_SECONDS)))
+            if (timestamp < now.AddSeconds(-1 * TOLERANCE_IN_SECONDS))
             {
                 throw new WebhookVerificationException("Message timestamp too old");
             }
-            if (timestamp > (now.AddSeconds(TOLERANCE_IN_SECONDS)))
+            if (timestamp > now.AddSeconds(TOLERANCE_IN_SECONDS))
             {
                 throw new WebhookVerificationException("Message timestamp too new");
             }
             return timestamp;
         }
-
-        public string Sign(string msgId, DateTimeOffset timestamp, string payload)
+        
+        public string Sign(ReadOnlySpan<char> msgId, DateTimeOffset timestamp, ReadOnlySpan<char> payload)
         {
-            var toSign = $"{msgId}.{timestamp.ToUnixTimeSeconds().ToString()}.{payload}";
-            var toSignBytes = SafeUTF8Encoding.GetBytes(toSign);
+            Span<char> signature = stackalloc char[SIGNATURE_LENGTH_STRING];
+            signature[0] = 'v';
+            signature[1] = '1';
+            signature[2] = ',';
+            CalculateSignature(msgId, timestamp, payload, signature.Slice(3), out var charsWritten);
+            return signature.Slice(0, charsWritten + 3).ToString();
+        }
+
+        private void CalculateSignature(
+            ReadOnlySpan<char> msgId, 
+            DateTimeOffset timestamp, 
+            ReadOnlySpan<char> payload, 
+            Span<char> signature,
+            out int charsWritten)
+        {
+            var timestampUnix = timestamp.ToUnixTimeSeconds().ToString();
+            
+            // Estimate buffer size and use stackalloc for smaller allocations
+            int msgIdLength = SafeUTF8Encoding.GetByteCount(msgId);
+            int payloadLength = SafeUTF8Encoding.GetByteCount(payload);
+            int timestampUnixLength = SafeUTF8Encoding.GetByteCount(timestampUnix);
+            
+            Span<byte> toSignBytes = stackalloc byte[msgIdLength + 1 + timestampUnixLength + 1 + payloadLength];
+            
+            SafeUTF8Encoding.GetBytes(msgId, toSignBytes.Slice(0, msgIdLength));
+            toSignBytes[msgIdLength] = (byte)'.';
+            SafeUTF8Encoding.GetBytes(timestampUnix, toSignBytes.Slice(msgIdLength + 1, timestampUnixLength));
+            toSignBytes[msgIdLength + 1 + timestampUnixLength] = (byte)'.';
+            SafeUTF8Encoding.GetBytes(payload, toSignBytes.Slice(msgIdLength + 1 + timestampUnixLength + 1));
+
+            Span<byte> signatureBin = stackalloc byte[SIGNATURE_LENGTH_BYTES];
+            CalculateSignature(toSignBytes, signatureBin);
+            
+            Span<byte> signatureB64 = stackalloc byte[SIGNATURE_LENGTH_BASE64];
+            var result = Base64.EncodeToUtf8(signatureBin, signatureB64, out _, out var bytesWritten);
+            if (result != OperationStatus.Done)
+                throw new WebhookVerificationException("Failed to encode signature to base64");
+            
+            if (!SafeUTF8Encoding.TryGetChars(signatureB64.Slice(0, bytesWritten), signature, out charsWritten))
+                throw new WebhookVerificationException("Failed to convert signature to utf8");
+        }
+        
+        private void CalculateSignature(ReadOnlySpan<byte> input, Span<byte> output)
+        {
             using (var hmac = new HMACSHA256(this.key))
             {
-                var hash = hmac.ComputeHash(toSignBytes);
-                var signature = Convert.ToBase64String(hash);
-                return $"v1,{signature}";
+                if (!hmac.TryComputeHash(input, output, out _))
+                    throw new WebhookVerificationException("Output buffer too small");
             }
         }
     }