Using latest csharp and dotnet features to increase performance and reduce memory footprint when creating and verifying webhook signatures

<!--
Thank you for your Pull Request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests.
-->

## Motivation

When using the webhook verification process we see lots of memory wasted and garbage collector kicking when our applications gets hit with a lot of webhook requests.

## Solution

* using ReadOnly<char> instead of string in Webhook.Verify and Webhook.Sign
* using stack alloctions instead of heap allocations when creating and verifing signatures as much as possible
* using new Base64 API to convert to base64 operating on spans


## Benchmarks

```csharp
[MemoryDiagnoser]
public class Tests
{
    private readonly string _msgId = "msg_p5jXN8AQM9LWM0D4loKWxJek";
    private readonly DateTimeOffset _timestamp = DateTimeOffset.FromUnixTimeSeconds(1614265330);
    private readonly string _payload = "{\"test\": 2432232314}";
    private readonly string _largePayload = Enumerable.Range(0, 1000).Select(i => i % 10).Aggregate(new StringBuilder(), (sb, i) => sb.Append(i), sb => sb.ToString());
    private readonly Webhook _webhook = new Webhook("whsec_MfKQ9r8GKYqrTwjUPD8ILPZIo2LaLaSw");
    private readonly WebHeaderCollection _headers;
    private readonly WebHeaderCollection _largeHeaders;
    
    public Tests()
    {
        var now = DateTimeOffset.UtcNow;
        _headers = new WebHeaderCollection
        {
            {"svix-id", _msgId},
            {"svix-timestamp", now.ToUnixTimeSeconds().ToString() },
            {"svix-signature", _webhook.Sign(_msgId, now, _payload)}
        };
        _largeHeaders = new WebHeaderCollection
        {
            {"svix-id", _msgId},
            {"svix-timestamp", now.ToUnixTimeSeconds().ToString() },
            {"svix-signature", _webhook.Sign(_msgId, now, _largePayload)}
        };
    }
    
    [Benchmark]
    public string? Sign()
    {
        return _webhook.Sign(_msgId, _timestamp, _payload);
    }
    
    [Benchmark]
    public string? SignLargePayload()
    {
        return _webhook.Sign(_msgId, _timestamp, _largePayload);
    }
    
    [Benchmark]
    public void Verify()
    {
        _webhook.Verify(_payload, _headers);
    }
    
    [Benchmark]
    public void VerifyLargePayload()
    {
        _webhook.Verify(_largePayload, _largeHeaders);
    }
}
```

Original Code


```
| Method             | Mean       | Error    | StdDev   | Gen0   | Allocated |
|------------------- |-----------:|---------:|---------:|-------:|----------:|
| Sign               |   945.0 ns |  3.75 ns |  3.13 ns | 0.1469 |     928 B |
| SignLargePayload   | 1,568.2 ns | 15.73 ns | 13.94 ns | 0.6142 |    3864 B |
| Verify             | 1,158.8 ns |  4.16 ns |  3.69 ns | 0.2117 |    1328 B |
| VerifyLargePayload | 1,772.6 ns |  5.76 ns |  4.49 ns | 0.6790 |    4264 B |
```

Code from this PR

```
| Method             | Mean       | Error   | StdDev  | Gen0   | Allocated |
|------------------- |-----------:|--------:|--------:|-------:|----------:|
| Sign               |   466.2 ns | 0.68 ns | 0.57 ns | 0.0267 |     168 B |
| SignLargePayload   |   885.4 ns | 3.20 ns | 2.68 ns | 0.0267 |     168 B |
| Verify             |   705.6 ns | 2.24 ns | 1.98 ns | 0.0095 |      64 B |
| VerifyLargePayload | 1,138.2 ns | 3.26 ns | 2.72 ns | 0.0095 |      64 B |
```