@esskar 2025-01-05
> Hey, thanks for the contribution, and sorry to hear the current code was thrashing the GC.
> 
> I made some comments about specific changes (not all of them yet), though would you mind opening smaller PRs that are more focused on just one change so that things are easier to review?
> 
> At the moment there are 4+ completely separate changes here which make it very hard to: (1) understand the reasoning behind each change / need, and (2) understanding what's going on so that we can review it effectively.
> 
> Thanks a lot!

Will do. I will also make some comments to your questions here.